Add/Update users and their type to our database when they login
This uses the NextAuth.js `signIn` event to `upsert` into our database
when users login.

This means we'll always be in sync with the Keycloak database as we
fetch the data whenever a user logs in.

If their role has changed since last login, we'll also automatically
update this in our database, because we're using `upsert` to update if
exists.

There were various methods of doing this. One option was to create a
Keycloak event listener that listens for user registrations/logins and
updates our database manually. I decided agains that route because it
involved creating a java package which would further complicate our
codebase and add yet another 'depdendency' which would make it difficult
to deploy and update.

Using the NextAuth.js approach should be safe because we're definitely
going to be using this auth endpoint to login going forward, so there
shouldn't be any chance of being out of sync with the Keycloak database.