Remove user_type tables
Now that we're using Keycloak, we can store all the user-related
attributes as User Attributes in the Keycloak database. This has the
benefits of removing possible duplication between the two databases, and
being able to completely separate user-specific data into the Keycloak
database, meaning we can make full use of the Keycloak 'abstraction' in
our system.

If we use a different login provider in future, we can just configure
Keycloak to use it as an Identity Provider, but still use Keycloak's
database for user-specific attributes.

In practice, all this change means is that we will store the department,
hospital, and health board IDs for each user inside Keycloak itself.