Add React Suite dependency
We might be able to replace Material UI with this. It seems to have a
more developer friendly experience and has a wider range of free
open-source components, such as a Date Range Picker which might be
useful for us!

The design system itself is quite similar to Material UI too!