Rework method of overrides for GET /api/responses, inc. hospitals
This method might be a bit easier to follow: if there are user_id,
department_id, or hospital_id query parameters, and the logged in user
has the potential to use one of these overrides, then it will be
applied. If none of the supplied overrides match the logged in user's
allowed overrides, the default behaviour will happen instead.