Remove ability to change the standard of a question
If we allow users to change the standard tied to a question, we'll end
up having different users answer the same question, but having their
scores stored against a different standard. When aggregating the data,
this will skew the results and make them less meaningful, so we
shouldn't let admins change the standard for a question.