Fix Join Code registration flow
We're now going to fetch the user profile every time the session is
queried, so we will always have the most up to date information.

For some reason, NextAuth.js only calls the jwt() callback with the
user/account/profile/isNewUser parameters on first login, and then never
again, which is why the session wasn't being updated.

Now we shouldn't need to rely on a forced logout to update the session,
improving the UX (and fixing the bug).