Fix profile fetching logic
I was relying on the session token always being valid previously, but
it's actually only valid from around 5 minutes from time of creation.

We're now going to use the refresh token too, and if the access token
doesn't work, attempt to regenerate the access token using the refresh
token.

We'll also store the newly-generated access token/refresh token if we
can, so we don't keep requesting a new token unneccessarily.