Return null session if profile fetch fails; Add more logging
If we can't fetch the profile, instead of returning a session object
with UNKNOWN_USER, if we return null then the user should be shown an
unregistered view, letting them log in and fix the issue.

Otherwise, there isn't any way for this error to recover unless the user
logs out themselves.