Set up API Jest test environment
This uses the node-pg module to setup the database schema before all
tests run, adding the bare tables and seeding some basic data (like
standards, questions, etc.).

Main issue here is that we need to override the DATABASE_URL environment
variable before the tests are run -- we can't do it within the code.
This is probably going to have to be part of the package.json script
using cross-env perhaps.