Create initial test for /api/recent_words
Lots of issues to get here:

- jest-openapi doesn't support node-fetch's Response types/format, which
next-test-api-route-handler uses internally
- jest-openapi requires the request paths to match the path in the spec
file, but that is the production URL not the localhost URL so will never
work without hackily updating the spec
- jest-openapi internally uses openapi-response-validator, which we are
now using to do the response validation
- we're now mocking NextAuth.js' useSession and getSession to mock the
session existence, so we don't need to mess around with Keycloak to let
our tests work. Previously I was trying to figure out how to
authenticate the user against Keycloak, get an access token, get a
session token, and add the session token to all requests, but if this
works that's great because we don't really need to test Keycloak --
that's the reason we're using a 3rd party option which is already
tested. The entire system testing will be done with our Puppeteer
end-to-end testing anyway, so let's make these API integration tests as
narrow as possible.