Also deploy to gh-pages branch
We need to do this because this repository is private, so we can't
access the care-quality-dashboard-api.json file unless we're
authenticated and have permission to view the repo, so it wouldn't work
for most people to view the rendered Swagger docs.

If/when this repo becomes public, we should be able to revert this
commit, and update the gh-pages/backend/index.html file to point to the
JSON file in this main branch itself.