Dereference OpenAPI schema for tests
openapi-response-validator doesn't follow references 100% correctly, so
wasn't following our response $ref's correctly (it strips out the $ref
so requires the response body to be inlined for every endpoint).

The best way to circumvent this is to dereference the entire schema, so
everything is inlined, making it a plain JSON object with no special
OpenAPI/JSON Schema syntax.