Create E2E Jest Test Environment
This will reset the `test` database for every test suite, and re-create
the Keycloak realm with new users again for each test suite.

This means each test suite will have a fresh start.

A few default users are also added to the Keycloak instance in this
setup script: clinician, department, hospital, health board, and admin
users. We can update these as and when we need.