Fix E2E test environment Keycloak Roles configuration
Keycloak docs say it is possible to create a new user and assign their
roles at the same time using the realmRoles property.

But that doesn't work at all.

So now we're doing it in 2 steps: create the user with the basic
information, attributes, credentials, and THEN configure their role
separately. It gets a bit messy but I've separated the parts as much as
I can!