Fix end-to-end test framework
We had a weird error where running multiple test suites serially failed
when they both use the database.

We were getting errors like "cache lookup failed for type x" and "cached
plan must not change result type".

The error seems to be something to do with the way Prisma is querying
the underlying database but I haven't been able to figure this out --
Prisma doesn't even cache prepared statements at the moment, so errors
to do with the fact that we're recreating the database schema etc.
shouldn't occur because there should already be no cache!

Running the queries directly on the PostgreSQL database (via psql) works
fine which is why I think it's something to do with Prisma's
interaction.

I couldn't fix the root cause, but we're now **restarting** the Next
(web-app) server for each test suite (file), which causes the error to
go away (maybe it was some Node caching issue? but again, it shouldn't
have been a problem because we were even prisma.$disconnect()ing for
each test(-suite) so the new schema should have been taken into
account...).

We're also no longer using jest-puppeteer because there seemed to be
some conflict with how we are now setting a global variable, and how it
sets its own global variables for `page`, `browser`, etc. (maybe
individual test environments override the entire global namespace?) --
so now we're using our own test environment to create our own Browser
and Page instances that can be used throughout the tests.

A side effect of this is that it might be easier to actually use
different puppeteer drivers for different browsers now, e.g. Edge
perhaps?