Fix WordsQuestions state handling
We were assuming that the setWordX caused an immediate update to the
state, but React actually queues the state updates for efficiency
(prevent unneccessary renders).

We need to add a useEffect() hook to tell the parent about the new words
whenever React decides to perform the re-render with the new state
instead.