Fix Docz production frontend docs
Using React Suite's `Panel` component (and others probably) was causing
a `TypeError: Cannot convert undefined or null to object` error when the
Docz site was built for production (but not in dev).

It looks like it was because of the
https://github.com/oliviertassinari/babel-plugin-transform-react-remove-prop-types
plugin which is enabled by default in Gatsby, which removes PropTypes
from code in production (but not in development). React Suite's code
requires the PropTypes to be there, so we were getting this error.

This change disables transforming for node_modules entirely:
`https://github.com/gatsbyjs/gatsby/issues/25297#issuecomment-651163951`.