Fix Docz production build
Gatsby builds it for the server as a static site, so uses server-side
rendering, but React Suite Alert requires the browser
`window`/`document` objects so the build was failing.

This will now override the React Suite Alert with a dummy stripped down
version that does nothing to fiz server-side renderingbuild issues.

This is similar to what Gatsby officially recommends doing:
https://www.gatsbyjs.com/docs/debugging-html-builds/#fixing-third-party-modules.